+++ src/blas/matrix/copy/d2f.sh
@@ -1,3 +1,0 @@
-cat double_copy.cpp \
-    | sed -e 's/double/float/g' \
-    > float_copy.cpp