+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Unreleased
 - add is_same_device_mem_stat https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/226
 
 ### Changed
+- change matrix BLAS template (without subvec_op and matmul) https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/238
 - change vector BLAS template https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/237
 - change vecadd/vecsub template https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/236
 - change LOBPCG eigensolver using xpay, axpyz instead of scal, vecadd, and vecsub https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/230